rename <SearchResults/> to <Search/>
